+++ gcs/src/test/java/com/google/cloud/hadoop/fs/gcs/GoogleHadoopFileSystemTest.java
@@ -371,4 +371,7 @@ public class GoogleHadoopFileSystemTest extends GoogleHadoopFileSystemIntegratio
 
   @Override
   public void testMd5FileChecksum() {}
+
+  @Override
+  public void testConcurrentCreationWithoutOverwrite_onlyOneSucceeds() {}
 }