+++ gcsio/src/test/java/com/google/cloud/hadoop/gcsio/GoogleCloudStorageTest.java
@@ -182,7 +182,6 @@ public class GoogleCloudStorageTest {
 
     MockHttpTransport transport =
         mockTransport(
-            jsonErrorResponse(ErrorResponses.NOT_FOUND),
             resumableUploadResponse(BUCKET_NAME, OBJECT_NAME),
             jsonDataResponse(
                 newStorageObject(BUCKET_NAME, OBJECT_NAME)
@@ -198,13 +197,15 @@ public class GoogleCloudStorageTest {
 
     assertThat(trackingHttpRequestInitializer.getAllRequestStrings())
         .containsExactly(
-            getRequestString(BUCKET_NAME, OBJECT_NAME),
             resumableUploadRequestString(
-                BUCKET_NAME, OBJECT_NAME, /* generationId= */ 0, /* replaceGenerationId= */ false),
+                BUCKET_NAME,
+                OBJECT_NAME,
+                /* generationId= */ null,
+                /* replaceGenerationId= */ false),
             resumableUploadChunkRequestString(BUCKET_NAME, OBJECT_NAME, /* uploadId= */ 1))
         .inOrder();
 
-    HttpRequest chunkUploadRequest = trackingHttpRequestInitializer.getAllRequests().get(2);
+    HttpRequest chunkUploadRequest = trackingHttpRequestInitializer.getAllRequests().get(1);
     assertThat(chunkUploadRequest.getContent().getLength()).isEqualTo(testData.length);
     try (ByteArrayOutputStream writtenData = new ByteArrayOutputStream(testData.length)) {
       chunkUploadRequest.getContent().writeTo(writtenData);
@@ -270,9 +271,11 @@ public class GoogleCloudStorageTest {
     assertThat(thrown).hasMessageThat().isEqualTo("Upload failed for 'gs://foo-bucket/bar-object'");
     assertThat(trackingHttpRequestInitializer.getAllRequestStrings())
         .containsExactly(
-            getRequestString(BUCKET_NAME, OBJECT_NAME),
             resumableUploadRequestString(
-                BUCKET_NAME, OBJECT_NAME, /* generationId= */ 1, /* replaceGenerationId= */ true))
+                BUCKET_NAME,
+                OBJECT_NAME,
+                /* generationId= */ null,
+                /* replaceGenerationId= */ true))
         .inOrder();
   }
 
@@ -285,7 +288,6 @@ public class GoogleCloudStorageTest {
 
     MockHttpTransport transport =
         mockTransport(
-            emptyResponse(HttpStatusCodes.STATUS_CODE_NOT_FOUND),
             resumableUploadResponse(BUCKET_NAME, OBJECT_NAME),
             jsonErrorResponse(ErrorResponses.GONE),
             resumableUploadResponse(BUCKET_NAME, OBJECT_NAME),
@@ -309,16 +311,21 @@ public class GoogleCloudStorageTest {
 
     assertThat(trackingHttpRequestInitializer.getAllRequestStrings())
         .containsExactly(
-            getRequestString(BUCKET_NAME, OBJECT_NAME),
             resumableUploadRequestString(
-                BUCKET_NAME, OBJECT_NAME, /* generationId= */ 0, /* replaceGenerationId= */ false),
+                BUCKET_NAME,
+                OBJECT_NAME,
+                /* generationId= */ null,
+                /* replaceGenerationId= */ false),
             resumableUploadChunkRequestString(BUCKET_NAME, OBJECT_NAME, /* uploadId= */ 1),
             resumableUploadRequestString(
-                BUCKET_NAME, OBJECT_NAME, /* generationId= */ 0, /* replaceGenerationId= */ false),
+                BUCKET_NAME,
+                OBJECT_NAME,
+                /* generationId= */ null,
+                /* replaceGenerationId= */ false),
             resumableUploadChunkRequestString(BUCKET_NAME, OBJECT_NAME, /* uploadId= */ 2))
         .inOrder();
 
-    HttpRequest writeRequest = trackingHttpRequestInitializer.getAllRequests().get(4);
+    HttpRequest writeRequest = trackingHttpRequestInitializer.getAllRequests().get(3);
     assertThat(writeRequest.getContent().getLength()).isEqualTo(testData.length);
     try (ByteArrayOutputStream writtenData = new ByteArrayOutputStream(testData.length)) {
       writeRequest.getContent().writeTo(writtenData);
@@ -335,7 +342,6 @@ public class GoogleCloudStorageTest {
 
     MockHttpTransport transport =
         mockTransport(
-            emptyResponse(HttpStatusCodes.STATUS_CODE_NOT_FOUND),
             resumableUploadResponse(BUCKET_NAME, OBJECT_NAME),
             // "308 Resume Incomplete" - successfully uploaded 1st chunk
             emptyResponse(308).addHeader("Range", "bytes=0-" + (uploadChunkSize - 1)),
@@ -363,20 +369,25 @@ public class GoogleCloudStorageTest {
 
     assertThat(trackingHttpRequestInitializer.getAllRequestStrings())
         .containsExactly(
-            getRequestString(BUCKET_NAME, OBJECT_NAME),
             resumableUploadRequestString(
-                BUCKET_NAME, OBJECT_NAME, /* generationId= */ 0, /* replaceGenerationId= */ false),
+                BUCKET_NAME,
+                OBJECT_NAME,
+                /* generationId= */ null,
+                /* replaceGenerationId= */ false),
             resumableUploadChunkRequestString(BUCKET_NAME, OBJECT_NAME, /* uploadId= */ 1),
             resumableUploadChunkRequestString(BUCKET_NAME, OBJECT_NAME, /* uploadId= */ 2),
             resumableUploadRequestString(
-                BUCKET_NAME, OBJECT_NAME, /* generationId= */ 0, /* replaceGenerationId= */ false),
+                BUCKET_NAME,
+                OBJECT_NAME,
+                /* generationId= */ null,
+                /* replaceGenerationId= */ false),
             resumableUploadChunkRequestString(BUCKET_NAME, OBJECT_NAME, /* uploadId= */ 3),
             resumableUploadChunkRequestString(BUCKET_NAME, OBJECT_NAME, /* uploadId= */ 4))
         .inOrder();
 
-    HttpRequest writeRequestChunk1 = trackingHttpRequestInitializer.getAllRequests().get(5);
+    HttpRequest writeRequestChunk1 = trackingHttpRequestInitializer.getAllRequests().get(4);
     assertThat(writeRequestChunk1.getContent().getLength()).isEqualTo(testData.length / 2);
-    HttpRequest writeRequestChunk2 = trackingHttpRequestInitializer.getAllRequests().get(6);
+    HttpRequest writeRequestChunk2 = trackingHttpRequestInitializer.getAllRequests().get(5);
     assertThat(writeRequestChunk2.getContent().getLength()).isEqualTo(testData.length / 2);
     try (ByteArrayOutputStream writtenData = new ByteArrayOutputStream(testData.length)) {
       writeRequestChunk1.getContent().writeTo(writtenData);
@@ -394,7 +405,6 @@ public class GoogleCloudStorageTest {
 
     MockHttpTransport transport =
         mockTransport(
-            emptyResponse(HttpStatusCodes.STATUS_CODE_NOT_FOUND),
             resumableUploadResponse(BUCKET_NAME, OBJECT_NAME),
             jsonErrorResponse(ErrorResponses.GONE),
             resumableUploadResponse(BUCKET_NAME, OBJECT_NAME),
@@ -419,16 +429,21 @@ public class GoogleCloudStorageTest {
 
     assertThat(trackingHttpRequestInitializer.getAllRequestStrings())
         .containsExactly(
-            getRequestString(BUCKET_NAME, OBJECT_NAME),
             resumableUploadRequestString(
-                BUCKET_NAME, OBJECT_NAME, /* generationId= */ 0, /* replaceGenerationId= */ false),
+                BUCKET_NAME,
+                OBJECT_NAME,
+                /* generationId= */ null,
+                /* replaceGenerationId= */ false),
             resumableUploadChunkRequestString(BUCKET_NAME, OBJECT_NAME, /* uploadId= */ 1),
             resumableUploadRequestString(
-                BUCKET_NAME, OBJECT_NAME, /* generationId= */ 0, /* replaceGenerationId= */ false),
+                BUCKET_NAME,
+                OBJECT_NAME,
+                /* generationId= */ null,
+                /* replaceGenerationId= */ false),
             resumableUploadChunkRequestString(BUCKET_NAME, OBJECT_NAME, /* uploadId= */ 2))
         .inOrder();
 
-    HttpRequest writeRequest = trackingHttpRequestInitializer.getAllRequests().get(4);
+    HttpRequest writeRequest = trackingHttpRequestInitializer.getAllRequests().get(3);
     assertThat(writeRequest.getContent().getLength()).isEqualTo(2 * testData.length);
     try (ByteArrayOutputStream writtenData = new ByteArrayOutputStream(testData.length)) {
       writeRequest.getContent().writeTo(writtenData);
@@ -445,7 +460,6 @@ public class GoogleCloudStorageTest {
 
     MockHttpTransport transport =
         mockTransport(
-            emptyResponse(HttpStatusCodes.STATUS_CODE_NOT_FOUND),
             resumableUploadResponse(BUCKET_NAME, OBJECT_NAME),
             // "308 Resume Incomplete" - successfully uploaded 1st chunk
             emptyResponse(308).addHeader("Range", "bytes=0-" + (uploadChunkSize - 1)),
@@ -476,26 +490,31 @@ public class GoogleCloudStorageTest {
 
     assertThat(trackingHttpRequestInitializer.getAllRequestStrings())
         .containsExactly(
-            getRequestString(BUCKET_NAME, OBJECT_NAME),
             resumableUploadRequestString(
-                BUCKET_NAME, OBJECT_NAME, /* generationId= */ 0, /* replaceGenerationId= */ false),
+                BUCKET_NAME,
+                OBJECT_NAME,
+                /* generationId= */ null,
+                /* replaceGenerationId= */ false),
             resumableUploadChunkRequestString(BUCKET_NAME, OBJECT_NAME, /* uploadId= */ 1),
             resumableUploadChunkRequestString(BUCKET_NAME, OBJECT_NAME, /* uploadId= */ 2),
             resumableUploadRequestString(
-                BUCKET_NAME, OBJECT_NAME, /* generationId= */ 0, /* replaceGenerationId= */ false),
+                BUCKET_NAME,
+                OBJECT_NAME,
+                /* generationId= */ null,
+                /* replaceGenerationId= */ false),
             resumableUploadChunkRequestString(BUCKET_NAME, OBJECT_NAME, /* uploadId= */ 3),
             resumableUploadChunkRequestString(BUCKET_NAME, OBJECT_NAME, /* uploadId= */ 4),
             resumableUploadChunkRequestString(BUCKET_NAME, OBJECT_NAME, /* uploadId= */ 5),
             resumableUploadChunkRequestString(BUCKET_NAME, OBJECT_NAME, /* uploadId= */ 6))
         .inOrder();
 
-    HttpRequest writeRequestChunk1 = trackingHttpRequestInitializer.getAllRequests().get(5);
+    HttpRequest writeRequestChunk1 = trackingHttpRequestInitializer.getAllRequests().get(4);
     assertThat(writeRequestChunk1.getContent().getLength()).isEqualTo(testData.length / 2);
-    HttpRequest writeRequestChunk2 = trackingHttpRequestInitializer.getAllRequests().get(6);
+    HttpRequest writeRequestChunk2 = trackingHttpRequestInitializer.getAllRequests().get(5);
     assertThat(writeRequestChunk2.getContent().getLength()).isEqualTo(testData.length / 2);
-    HttpRequest writeRequestChunk3 = trackingHttpRequestInitializer.getAllRequests().get(7);
+    HttpRequest writeRequestChunk3 = trackingHttpRequestInitializer.getAllRequests().get(6);
     assertThat(writeRequestChunk3.getContent().getLength()).isEqualTo(testData.length / 2);
-    HttpRequest writeRequestChunk4 = trackingHttpRequestInitializer.getAllRequests().get(8);
+    HttpRequest writeRequestChunk4 = trackingHttpRequestInitializer.getAllRequests().get(7);
     assertThat(writeRequestChunk4.getContent().getLength()).isEqualTo(testData.length / 2);
     try (ByteArrayOutputStream writtenData = new ByteArrayOutputStream(testData.length)) {
       writeRequestChunk1.getContent().writeTo(writtenData);
@@ -515,7 +534,6 @@ public class GoogleCloudStorageTest {
 
     MockHttpTransport transport =
         mockTransport(
-            emptyResponse(HttpStatusCodes.STATUS_CODE_NOT_FOUND),
             resumableUploadResponse(BUCKET_NAME, OBJECT_NAME),
             jsonErrorResponse(ErrorResponses.GONE));
 
@@ -2441,9 +2459,7 @@ public class GoogleCloudStorageTest {
     gcs.compose(BUCKET_NAME, sources, OBJECT_NAME, CreateFileOptions.DEFAULT_CONTENT_TYPE);
 
     assertThat(trackingHttpRequestInitializer.getAllRequestStrings())
-        .containsExactly(
-            getRequestString(BUCKET_NAME, OBJECT_NAME),
-            composeRequestString(BUCKET_NAME, OBJECT_NAME, 1))
+        .containsExactly(composeRequestString(BUCKET_NAME, OBJECT_NAME, /* generationId= */ null))
         .inOrder();
   }